% Generated by roxygen2 (4.0.1): do not edit by hand
\name{completeBetaCpp}
\alias{completeBetaCpp}
\title{Extends the parameter vector to the natural dimension}
\usage{
completeBetaCpp(beta, blockLength)
}
\arguments{
\item{beta}{the parameter vector}

\item{blockLength}{the vector of blockLengths corresponding to each entry in beta}
}
\value{
the extended parameter vector
}
\description{
Extends the parameter vector to the natural dimension
}

