% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convexMinorant}
\alias{convexMinorant}
\title{Computes the convex minorant of a polygon.}
\usage{
convexMinorant(y, x)
}
\value{
vector of the y-coordinates of the convex minorant
}
\description{
Computes the convex minorant of a polygon.
}

