% Generated by roxygen2 (4.0.1): do not edit by hand
\name{penaltyCpp}
\alias{penaltyCpp}
\title{Computes the penalty term for a given parameter vector}
\usage{
penaltyCpp(beta, lambda, method = "1order")
}
\arguments{
\item{beta}{parameter vector}

\item{lambda}{multiplicator}

\item{method}{(eiter "1order" = TV or "2order" = sum Second Derivative)}
}
\value{
penalty
}
\description{
Computes the penalty term for a given parameter vector
}

