% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rcpp_convex_minorant}
\alias{rcpp_convex_minorant}
\title{The length of a string (in characters).}
\usage{
rcpp_convex_minorant(x, y)
}
\arguments{
\item{str}{input character vector}
}
\value{
characters in each element of the vector
}
\description{
The length of a string (in characters).
}

