\name{rqIcm-package}
\alias{rqIcm-package}
\alias{rqIcm}
\docType{package}
\title{
Isotone quantile regression
}
\description{
Uses the iterative convex minorant algorithm for isotone quantile regression
}
\details{
\tabular{ll}{
Package: \tab rqIcm\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-06-10\cr
License: \tab What Licence is it under ?\cr
}
The most important function is rqIcm
}
\author{ M. Roth
Maintainer: Who to complain to <yourfault@somewhere.net>
}
\references{
Roth et al.
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
%%\keyword{ package }
%%\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%%}
%%\examples{
%% ~~ simple examples of the most important functions ~~
%%}
