% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rqSmoothRegional}
\alias{rqSmoothRegional}
\title{Computes the smoothed objective criterion for multiple samples}
\usage{
rqSmoothRegional(y, beta, blockLength, tau, lambda = 0, method = "1order",
  c = 0.1)
}
\arguments{
\item{y}{the data}

\item{beta}{the parameter vector}

\item{blockLength}{the blocklength vector}

\item{tau}{the quantile}
}
\description{
Computes the smoothed objective criterion for multiple samples
}

